requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0

# LLM Providers (install the ones you need)
openai>=1.0.0  # For OpenAI/GPT models
google-generativeai>=0.3.0  # For Google Gemini
anthropic>=0.18.0  # For Anthropic Claude

# Web search
tavily-python>=0.3.0

# Redis for session management
redis>=5.0.0
hiredis>=2.2.0  # C parser for better performance

# Optional: For production deployment
uvloop>=0.19.0  # Faster event loop
orjson>=3.9.0  # Faster JSON parsing

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0
ruff>=0.1.0